addressed memory issues, still getting nan issues with loss

+++ v2/helpers.py
@@ -5,7 +5,7 @@ def im2col(X, k_height, k_width, padding=1, stride=1):
 
     out_height = (H + 2*padding - k_height)//stride + 1
     out_width = (W + 2*padding - k_width)//stride + 1
-    out = np.zeros((k_height, k_width, N, C, out_height, out_width))
+    out = np.zeros((k_height, k_width, N, C, out_height, out_width), dtype=np.float32)
     
     padded_X = np.pad(X, [(0,0), (0,0), (padding, padding), (padding, padding)])
 
@@ -28,7 +28,7 @@ def im2col_bw(grad_X_col, X_shape, k_height, k_width, padding=1, stride=1):
     grad_reshape = np.reshape(grad_X_col, (C, k_height, k_width, out_height, out_width, N))
     grad_transpose = np.transpose(grad_reshape, (1, 2, 5, 0, 3, 4))
 
-    input = np.zeros((N, C, H + 2*padding, W + 2*padding))
+    input = np.zeros((N, C, H + 2*padding, W + 2*padding), dtype=np.float32)
     r_end = stride*out_height
     for r in range(k_height):
         c_end = stride*out_width
